Update swissqr_report.py
When we confirm an invoice, the payment_reference is created and is similar as name in account.move. It is not possible to generate the QR code for the invoice because the argument passed through this payment_reference should be an ISR number.

Therefor I suggest to pass l10n_ch_isr_number as the ISR argument instead of payment_reference.